requirements.txt
streamlit==0.78.0
pandas==1.2.3
plotly==4.14.3
numpy==1.20.1
opencv_python==4.5.1.48
matplotlib==3.3.4
scipy==1.6.2
Pillow==8.2.0
scikit_learn==0.24.1
